requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
firebase-admin==6.2.0
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
schedule==1.2.0 